Fix or silence several type conversion warnings in x64 builds.

+++ Sources/Plasma/Apps/plPythonPack/main.cpp
@@ -191,7 +191,7 @@ void WritePythonFile(const plFileName &fileName, const plFileName &path, hsStrea
         ST::printf(out, "\n");
 
         s->WriteLE32((int32_t)size);
-        s->Write(size, pycode);
+        s->Write((uint32_t)size, pycode);
         delete[] pycode;
     }
     else
